+++ sdk/python/tests/test_Network.py
@@ -1,26 +1,52 @@
+import datetime
 import unittest
 
 from symbolchain.Network import Network, NetworkLocator
+from symbolchain.NetworkTimestamp import NetworkTimestamp, NetworkTimestampDatetimeConverter
 
 
 class NetworkTest(unittest.TestCase):
+	def test_can_convert_network_time_to_datetime(self):
+		# Arrange:
+		converter = NetworkTimestampDatetimeConverter(datetime.datetime(2022, 3, 16, 0, 6, 25, tzinfo=datetime.timezone.utc), 'minutes')
+		network = Network('foo', 0x55, converter, None, None)
+
+		# Act:
+		datetime_timestamp = network.to_datetime(NetworkTimestamp(60))
+
+		# Assert:
+		self.assertEqual(datetime.datetime(2022, 3, 16, 1, 6, 25, tzinfo=datetime.timezone.utc), datetime_timestamp)
+
+	def test_can_convert_datetime_to_network_time(self):
+		# Arrange:
+		converter = NetworkTimestampDatetimeConverter(datetime.datetime(2022, 3, 16, 0, 6, 25, tzinfo=datetime.timezone.utc), 'minutes')
+		network = Network('foo', 0x55, converter, None, NetworkTimestamp)
+
+		# Act:
+		network_timestamp = network.from_datetime(datetime.datetime(2022, 3, 16, 1, 6, 25, tzinfo=datetime.timezone.utc))
+
+		# Assert:
+		self.assertEqual(60, network_timestamp.timestamp)
+
 	def test_equality_is_supported(self):
 		# Arrange:
-		network = Network('foo', 0x55, None)
+		network = Network('foo', 0x55, None, None, None)
 
 		# Act + Assert:
-		self.assertEqual(network, Network('foo', 0x55, None))
-		self.assertEqual(network, Network('foo', 0x55, int))  # address_class is not compared
+		self.assertEqual(network, Network('foo', 0x55, None, None, None))
+		self.assertEqual(network, Network('foo', 0x55, {}, None, None))  # datetime_converter is not compared
+		self.assertEqual(network, Network('foo', 0x55, None, int, None))  # address_class is not compared
+		self.assertEqual(network, Network('foo', 0x55, None, None, int))  # network_timestamp_class is not compared
 
-		self.assertNotEqual(network, Network('Foo', 0x55, None))
-		self.assertNotEqual(network, Network('foo', 0x54, None))
+		self.assertNotEqual(network, Network('Foo', 0x55, None, None, None))
+		self.assertNotEqual(network, Network('foo', 0x54, None, None, None))
 		self.assertNotEqual(network, None)
 
 	def test_string_is_supported(self):
-		self.assertEqual('foo', str(Network('foo', 0x55, None)))
+		self.assertEqual('foo', str(Network('foo', 0x55, None, None, None)))
 
 
-PREDEFINED_NETWORKS = [Network('foo', 0x55, None), Network('bar', 0x37, None)]
+PREDEFINED_NETWORKS = [Network('foo', 0x55, None, None, None), Network('bar', 0x37, None, None, None)]
 
 
 class NetworkLocatorTest(unittest.TestCase):