+++ sdk/javascript/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## next
+
+### Changed
+ - toDatetime and fromDatetime promoted from NetworkTimestamp to Network
+
 ## 3.0.0 - 02-Mar-2022
 
 ### Changed